Implementação da verificação da existência do sqlite e suas tabelas
Implementação da verificação da existência do sqlite e suas tabelas.

+++ Services/Interfaces/ITextToSpeakService.cs
@@ -4,6 +4,6 @@ namespace LetterStomach.Services.Interfaces
 {
     public interface ITextToSpeakService
     {
-        void SpeakText(List<Message> messages, string language, int pitch_speak, int volume_speak);
+        void SpeakText(List<Message> messages, string language, float pitch_speak, float volume_speak);
     }
 }
